+++ app/src/main/java/woowacourse/movie/domain/model/ScreenType.kt
@@ -1,0 +1,6 @@
+package woowacourse.movie.domain.model
+
+enum class ScreenType(val id: Int) {
+    SCREEN(0),
+    ADVERTISEMENT(1),
+}
