+++ app/src/main/java/woowacourse/movie/ui/screen/adapter/AdvertisementViewHolder.kt
@@ -4,6 +4,7 @@ import android.view.View
 import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
+import woowacourse.movie.domain.model.ScreenAd
 
 class AdViewHolder(view: View, private val onItemClick: (id: Int) -> Unit): RecyclerView.ViewHolder(view) {
     private val adImage: ImageView = view.findViewById(R.id.iv_advertisement)
