+++ app/src/main/java/woowacourse/movie/ui/screen/adapter/ScreenAdapter.kt
@@ -5,9 +5,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.domain.model.DateRange
-import woowacourse.movie.domain.model.DrawableImage
-import woowacourse.movie.ui.MoviePreviewUI
+import woowacourse.movie.domain.model.ScreenAd
+import woowacourse.movie.domain.model.ScreenType
 
 class ScreenAdapter(
     private val onItemClick: (id: Int) -> Unit,
@@ -17,6 +16,17 @@ class ScreenAdapter(
     override fun getItemViewType(position: Int): Int = when {
         ((position + 1) % ADVERTISEMENT_INTERVAL == 0) -> ScreenType.ADVERTISEMENT.id
         else -> ScreenType.SCREEN.id
+    }
+
+    override fun getItemCount(): Int {
+        val actualItemCount = super.getItemCount()
+        val additionalAds = actualItemCount / ADVERTISEMENT_INTERVAL
+        return actualItemCount + additionalAds
+    }
+
+    override fun getItem(position: Int): ScreenAd {
+        val adjustedPosition = position - (position / ADVERTISEMENT_INTERVAL)
+        return super.getItem(adjustedPosition)
     }
 
     override fun onCreateViewHolder(
@@ -41,77 +51,17 @@ class ScreenAdapter(
         }
     }
 
-    override fun getItemCount(): Int {
-        val actualItemCount = super.getItemCount()
-        val additionalAds = actualItemCount / ADVERTISEMENT_INTERVAL
-        return actualItemCount + additionalAds
-    }
-
-    override fun getItem(position: Int): ScreenAd {
-        val adjustedPosition = position - (position / ADVERTISEMENT_INTERVAL)
-        return super.getItem(adjustedPosition)
-
-    }
-
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         val item = getItem(position)
 
-        if (item is ScreenAd.ScreenPreviewUi2 && holder is ScreenViewHolder) {
+        if (item is ScreenAd.ScreenPreviewUi && holder is ScreenViewHolder) {
             holder.bind(item)
         } else if (item is ScreenAd.Advertisement && holder is AdViewHolder) {
             holder.bind(item)
         }
     }
 
-
     companion object {
         private const val ADVERTISEMENT_INTERVAL = 4
-    }
-}
-
-enum class ScreenType(val id: Int) {
-    SCREEN(0),
-    ADVERTISEMENT(1),
-}
-
-sealed interface ScreenAd {
-    val id: Int
-
-    fun equalsWith(other: ScreenAd?): Boolean
-
-    data class ScreenPreviewUi2(
-        override val id: Int,
-        val moviePreviewUI: MoviePreviewUI,
-        val dateRange: DateRange,
-    ) : ScreenAd {
-        override fun equalsWith(other: ScreenAd?): Boolean {
-            if (this === other) return true
-            if (javaClass != other?.javaClass) return false
-
-            other as ScreenPreviewUi2
-
-            if (id != other.id) return false
-            if (moviePreviewUI != other.moviePreviewUI) return false
-            if (dateRange != other.dateRange) return false
-
-            return true
-        }
-    }
-
-    data class Advertisement(
-        override val id: Int = 0,
-        val advertisement: DrawableImage,
-    ) : ScreenAd {
-        override fun equalsWith(other: ScreenAd?): Boolean {
-            if (this === other) return true
-            if (javaClass != other?.javaClass) return false
-
-            other as Advertisement
-
-            if (id != other.id) return false
-            if (advertisement != other.advertisement) return false
-
-            return true
-        }
     }
 }
