+++ app/src/main/java/woowacourse/movie/ui/screen/adapter/ScreenPreviewUiDiffUtil.kt
@@ -1,6 +1,7 @@
 package woowacourse.movie.ui.screen.adapter
 
 import androidx.recyclerview.widget.DiffUtil
+import woowacourse.movie.domain.model.ScreenAd
 
 class ScreenPreviewUiDiffUtil : DiffUtil.ItemCallback<ScreenAd>() {
     override fun areItemsTheSame(oldItem: ScreenAd, newItem: ScreenAd): Boolean {
